requirements.txt
aiohttp==3.8.3
aiohttp-retry==2.9.1
aiosignal==1.3.1
annotated-types==0.7.0
anvil-uplink==0.5.2
anyio==4.9.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
dataclasses-json==0.6.7
debugpy==1.6.5
distro==1.9.0
exceptiongroup==1.2.2
faiss-cpu==1.11.0
Flask==2.2.2
frozenlist==1.3.3
future==1.0.0
greenlet==3.2.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
httpx-sse==0.4.0
idna==3.4
importlib_metadata==8.7.0
iso8601==1.1.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.58
langchain-text-splitters==0.3.8
langsmith==0.3.42
MarkupSafe==2.1.2
marshmallow==3.26.1
multidict==6.0.4
mypy_extensions==1.1.0
numpy==1.26.4
openai==1.77.0
orjson==3.10.18
packaging==24.2
parso==0.8.3
passlib==1.7.4
pdfrw==0.4
pillow==11.2.1
platformdirs==2.6.2
pluggy==1.0.0
protobuf==4.21.12
pycparser==2.21
pycryptodome==3.23.0
pycryptodomex==3.16.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyflakes==2.5.0
PyMuPDF==1.26.0
pypdf==5.6.0
PyPDF2==3.0.1
pyseto==1.7.0
python-dotenv==1.1.0
python-lsp-jsonrpc==1.0.0
python-telegram-bot==20.3
pytoolconfig==1.2.5
PyYAML==6.0.2
regex==2024.11.6
replit==3.6.3
replit-python-lsp-server==1.15.9
reportlab==4.4.1
requests==2.28.2
requests-toolbelt==1.0.0
rope==1.7.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
telegram==0.0.1
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tomli==2.0.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
ujson==5.7.0
urllib3==1.26.14
Werkzeug==2.2.2
whatthepatch==1.0.3
ws4py-sslupdate==0.5.1b0
yapf==0.32.0
yarl==1.8.2
zipp==3.21.0
zstandard==0.23.0
